requirements.txt
django-vkontakte-api>=0.3.4